==> 2_Unimodal_MRI_CNN/3_Readlogs_Testset.ipynb <==
from _labels import *
from src._cnn_config import *
from pathlib import Path
import pandas as pd

brainparts = []
pathogens = []
val_foldlist = []
aurocs = []
f1s = []
accs = []
prcs = []

for part_to_segment in labels:
    part_to_segment = part_to_segment.replace('-', '_')
    partlogdir = homedir + '/TrainLogs/' + part_to_segment.upper() + 'LOGS'

    for pathogen in pathogenlist:
        pathogenlogdir = partlogdir + '/' + str(pathogen)

        for val_fold in range(num_val_folds):
                valfolddir = pathogenlogdir + '/val_fold_' + str(val_fold)

                val_best_log_path = Path(valfolddir + '/test_log.csv')
                df_val = pd.read_csv(val_best_log_path)
                try:
                    auroc = float(df_val.loc[0, 'roc_auc'])
                except:
                    print('No AUROC for ' + part_to_segment)
                    break

                acc = float(df_val.loc[0, 'acc'])

                val_best_confusion_log_path = Path(valfolddir + '/test_confusion_log.csv')
                df_valconf = pd.read_csv(val_best_confusion_log_path)

                TP = int(df_valconf.loc[0, 'TP'])
                FN = int(df_valconf.loc[0, 'FN'])
                FP = int(df_valconf.loc[0, 'FP'])
                
                prc = TP/(TP+FP) if (TP + FP) != 0 else 0
                f1 = TP/(TP + (FP + FN)/2) if (TP + (FP + FN)/2) != 0 else 0
                
                brainparts.append(part_to_segment)
                pathogens.append(pathogen)
                val_foldlist.append(val_fold)
                aurocs.append(auroc)
                f1s.append(f1)
                accs.append(acc)
                prcs.append(prc)

df_test_metrics = pd.DataFrame({'Brainpart': brainparts, 'Pathogen': pathogens, 'Test_Fold': val_foldlist, 'AUROC': aurocs, 'F1_Score': f1s, 'Acc': accs, 'Prc': prcs})

# Display Dataframe
df_test_metrics.head()

df_test_best_metrics = result = df_test_metrics.loc[df_test_metrics.groupby(['Brainpart', 'Pathogen'])['AUROC'].idxmax()]

# Display Dataframe
df_test_best_metrics.head()

df_test_best_metrics.to_csv('test_best_metrics.csv', index=False)
