==> 2_Unimodal_MRI_CNN/4_Best_Models_Extraction.ipynb <==
import pandas as pd
from inference_lookup import *
# All labels
labels = [
# 'Lateral-Ventricle',  
# 'Inf-Lat-Vent',
'Cerebellum-White-Matter',
'Cerebellum-Cortex',
'Thalamus-Proper',
'Caudate',
'Putamen',
'Pallidum',
# '3rd-Ventricle',
# '4th-Ventricle',
'Brain-Stem',
'Hippocampus',
'Amygdala',
# 'CSF',
'Accumbens-area',
'VentralDC',
'Choroid-plexus',
'caudalanteriorcingulate',
'caudalmiddlefrontal',
'cuneus',
'entorhinal',
'fusiform',
'inferiorparietal',
'inferiortemporal', 
'isthmuscingulate',
'lateraloccipital',
'lateralorbitofrontal',
'lingual',
'medialorbitofrontal',
'middletemporal',
'parahippocampal',
'paracentral',
'parsopercularis',
'parsorbitalis',
'parstriangularis',
'pericalcarine',
'postcentral',
'posteriorcingulate',
'precentral',
'precuneus',
'rostralanteriorcingulate',
'rostralmiddlefrontal',
'superiorfrontal',
'superiorparietal',
'superiortemporal',
'supramarginal',
'transversetemporal',
'insula'
]
import shutil
pathogenlist = [0]

best_metrics = pd.read_csv('/home/yhchoi/FSTrain_0205/test_best_metrics.csv')
trynum = 6

for pathogen in pathogenlist : 
    for brain_part in labels:
        brain_part = brain_part.replace('-', '_')
        fold = best_metrics[(best_metrics['Pathogen'] == pathogen) & (best_metrics['Brainpart'] == brain_part)]['Test_Fold'].values[0]
        src = f'/home/yhchoi/FSTrain_0205/TrainLogs/{brain_part.upper()}LOGS/{pathogen}/val_fold_{fold}/bestmodel.pt'
        dest = f'/home/yhchoi/FSTrain_0205/Inference/cnn_all/{pathogen}/{brain_part}_bestmodel.pt'
        shutil.copy(src, dest)
        print(f'Copied {src} to {dest}')
